fix: should avoid through variables in inlined module
