+++ lib/ConcatenationScope.js
@@ -5,13 +5,15 @@
 
 "use strict";
 
+const {
+	DEFAULT_EXPORT,
+	NAMESPACE_OBJECT_EXPORT
+} = require("./util/concatenate");
+
 /** @typedef {import("./Module")} Module */
 
 const MODULE_REFERENCE_REGEXP =
 	/^__WEBPACK_MODULE_REFERENCE__(\d+)_([\da-f]+|ns)(_call)?(_directImport)?(?:_asiSafe(\d))?__$/;
-
-const DEFAULT_EXPORT = "__WEBPACK_DEFAULT_EXPORT__";
-const NAMESPACE_OBJECT_EXPORT = "__WEBPACK_NAMESPACE_OBJECT__";
 
 /**
  * @typedef {object} ExternalModuleInfo
