+++ lib/javascript/JavascriptModulesPlugin.js
@@ -31,13 +31,21 @@ const Template = require("../Template");
 const { last, someInIterable } = require("../util/IterableHelpers");
 const StringXor = require("../util/StringXor");
 const { compareModulesByIdentifier } = require("../util/comparators");
+const {
+	getPathInAst,
+	getAllReferences,
+	RESERVED_NAMES,
+	findNewName,
+	addScopeSymbols,
+	getUsedNamesInScopeInfo
+} = require("../util/concatenate");
 const createHash = require("../util/createHash");
-const { getPathInAst, getAllReferences } = require("../util/mergeScope");
 const nonNumericOnlyHash = require("../util/nonNumericOnlyHash");
 const { intersectRuntime } = require("../util/runtime");
 const JavascriptGenerator = require("./JavascriptGenerator");
 const JavascriptParser = require("./JavascriptParser");
 
+/** @typedef {import("eslint-scope").Reference} Reference */
 /** @typedef {import("eslint-scope").Scope} Scope */
 /** @typedef {import("eslint-scope").Variable} Variable */
 /** @typedef {import("webpack-sources").Source} Source */
@@ -1458,8 +1466,7 @@ class JavascriptModulesPlugin {
 		const renamedInlinedModules = new Map();
 		const { runtimeTemplate } = renderContext;
 
-		/** @typedef {{ source: Source, ast: any, variables: Set<Variable>, usedInNonInlined: Set<Variable>}} InlinedModulesInfo */
-		/** @type {Map<Module, InlinedModulesInfo>} */
+		/** @type {Map<Module, { source: Source, module: Module, ast: any, variables: Set<Variable>, through: Set<Reference>, usedInNonInlined: Set<Variable>, moduleScope: Scope }>} */
 		const inlinedModulesToInfo = new Map();
 		/** @type {Set<string>} */
 		const nonInlinedModuleThroughIdentifiers = new Set();
@@ -1487,14 +1494,17 @@ class JavascriptModulesPlugin {
 				ignoreEval: true
 			});
 
-			const globalScope = /** @type {Scope} */ (scopeManager.acquire(ast));
+			const globalScope = scopeManager.acquire(ast);
 			if (inlinedModules && inlinedModules.has(m)) {
 				const moduleScope = globalScope.childScopes[0];
 				inlinedModulesToInfo.set(m, {
 					source: moduleSource,
 					ast,
+					module: m,
 					variables: new Set(moduleScope.variables),
-					usedInNonInlined: new Set()
+					through: new Set(moduleScope.through),
+					usedInNonInlined: new Set(),
+					moduleScope
 				});
 			} else {
 				for (const ref of globalScope.through) {
@@ -1505,7 +1515,10 @@ class JavascriptModulesPlugin {
 
 		for (const [, { variables, usedInNonInlined }] of inlinedModulesToInfo) {
 			for (const variable of variables) {
-				if (nonInlinedModuleThroughIdentifiers.has(variable.name)) {
+				if (
+					nonInlinedModuleThroughIdentifiers.has(variable.name) ||
+					RESERVED_NAMES.has(variable.name)
+				) {
 					usedInNonInlined.add(variable);
 				}
 			}
@@ -1519,39 +1532,70 @@ class JavascriptModulesPlugin {
 				continue;
 			}
 
-			const usedNames = new Set(
-				Array.from(
-					/** @type {InlinedModulesInfo} */
-					(inlinedModulesToInfo.get(m)).variables
-				).map(v => v.name)
+			const info = inlinedModulesToInfo.get(m);
+			const allUsedNames = new Set(
+				Array.from(info.through, v => v.identifier.name)
 			);
 
 			for (const variable of usedInNonInlined) {
+				allUsedNames.add(variable.name);
+			}
+
+			for (const variable of info.variables) {
+				allUsedNames.add(variable.name);
 				const references = getAllReferences(variable);
 				const allIdentifiers = new Set(
 					references.map(r => r.identifier).concat(variable.identifiers)
 				);
 
-				const newName = this.findNewName(
-					variable.name,
-					usedNames,
-					m.readableIdentifier(runtimeTemplate.requestShortener)
+				const usedNamesInScopeInfo = new Map();
+				const ignoredScopes = new Set();
+
+				const name = variable.name;
+				const { usedNames, alreadyCheckedScopes } = getUsedNamesInScopeInfo(
+					usedNamesInScopeInfo,
+					info.module.identifier(),
+					name
 				);
-				usedNames.add(newName);
-				for (const identifier of allIdentifiers) {
-					const r = /** @type {Range} */ (identifier.range);
-					const path = getPathInAst(ast, identifier);
-					if (path && path.length > 1) {
-						const maybeProperty =
-							path[1].type === "AssignmentPattern" && path[1].left === path[0]
-								? path[2]
-								: path[1];
-						if (maybeProperty.type === "Property" && maybeProperty.shorthand) {
-							source.insert(r[1], `: ${newName}`);
-							continue;
-						}
+
+				if (allUsedNames.has(name) || usedNames.has(name)) {
+					const references = getAllReferences(variable);
+					for (const ref of references) {
+						addScopeSymbols(
+							ref.from,
+							usedNames,
+							alreadyCheckedScopes,
+							ignoredScopes
+						);
 					}
-					source.replace(r[0], r[1] - 1, newName);
+
+					const newName = findNewName(
+						variable.name,
+						allUsedNames,
+						usedNames,
+						m.readableIdentifier(runtimeTemplate.requestShortener)
+					);
+					allUsedNames.add(newName);
+					for (const identifier of allIdentifiers) {
+						const r = identifier.range;
+						const path = getPathInAst(ast, identifier);
+						if (path && path.length > 1) {
+							const maybeProperty =
+								path[1].type === "AssignmentPattern" && path[1].left === path[0]
+									? path[2]
+									: path[1];
+							if (
+								maybeProperty.type === "Property" &&
+								maybeProperty.shorthand
+							) {
+								source.insert(r[1], `: ${newName}`);
+								continue;
+							}
+						}
+						source.replace(r[0], r[1] - 1, newName);
+					}
+				} else {
+					allUsedNames.add(name);
 				}
 			}
 
@@ -1559,38 +1603,6 @@ class JavascriptModulesPlugin {
 		}
 
 		return renamedInlinedModules;
-	}
-
-	/**
-	 * @param {string} oldName oldName
-	 * @param {Set<string>} usedName usedName
-	 * @param {string} extraInfo extraInfo
-	 * @returns {string} extraInfo
-	 */
-	findNewName(oldName, usedName, extraInfo) {
-		let name = oldName;
-
-		// Remove uncool stuff
-		extraInfo = extraInfo.replace(
-			/\.+\/|(\/index)?\.([a-zA-Z0-9]{1,4})($|\s|\?)|\s*\+\s*\d+\s*modules/g,
-			""
-		);
-		const splittedInfo = extraInfo.split("/");
-		while (splittedInfo.length) {
-			name = splittedInfo.pop() + (name ? `_${name}` : "");
-			const nameIdent = Template.toIdentifier(name);
-			if (!usedName.has(nameIdent)) {
-				return nameIdent;
-			}
-		}
-
-		let i = 0;
-		let nameWithNumber = Template.toIdentifier(`${name}_${i}`);
-		while (usedName.has(nameWithNumber)) {
-			i++;
-			nameWithNumber = Template.toIdentifier(`${name}_${i}`);
-		}
-		return nameWithNumber;
 	}
 }
