+++ lib/optimize/ConcatenatedModule.js
@@ -24,10 +24,17 @@ const JavascriptParser = require("../javascript/JavascriptParser");
 const { equals } = require("../util/ArrayHelpers");
 const LazySet = require("../util/LazySet");
 const { concatComparators } = require("../util/comparators");
+const {
+	RESERVED_NAMES,
+	findNewName,
+	addScopeSymbols,
+	getAllReferences,
+	getPathInAst,
+	getUsedNamesInScopeInfo
+} = require("../util/concatenate");
 const createHash = require("../util/createHash");
 const { makePathsRelative } = require("../util/identifier");
 const makeSerializable = require("../util/makeSerializable");
-const { getAllReferences, getPathInAst } = require("../util/mergeScope");
 const propertyAccess = require("../util/propertyAccess");
 const { propertyName } = require("../util/propertyName");
 const {
@@ -74,6 +81,7 @@ const {
 /** @typedef {import("../serialization/ObjectMiddleware").ObjectDeserializerContext} ObjectDeserializerContext */
 /** @typedef {import("../util/Hash")} Hash */
 /** @typedef {typeof import("../util/Hash")} HashConstructor */
+/** @typedef {import("../util/concatenate").UsedNames} UsedNames */
 /** @typedef {import("../util/fs").InputFileSystem} InputFileSystem */
 /** @typedef {import("../util/runtime").RuntimeSpec} RuntimeSpec */
 
@@ -171,52 +179,13 @@ if (!ReferencerClass.prototype.PropertyDefinition) {
  * @property {ConcatenatedModuleInfo | ExternalModuleInfo} target
  */
 
-/** @typedef {Set<string>} UsedNames */
-
-const RESERVED_NAMES = new Set(
-	[
-		// internal names (should always be renamed)
-		ConcatenationScope.DEFAULT_EXPORT,
-		ConcatenationScope.NAMESPACE_OBJECT_EXPORT,
-
-		// keywords
-		"abstract,arguments,async,await,boolean,break,byte,case,catch,char,class,const,continue",
-		"debugger,default,delete,do,double,else,enum,eval,export,extends,false,final,finally,float",
-		"for,function,goto,if,implements,import,in,instanceof,int,interface,let,long,native,new,null",
-		"package,private,protected,public,return,short,static,super,switch,synchronized,this,throw",
-		"throws,transient,true,try,typeof,var,void,volatile,while,with,yield",
-
-		// commonjs/amd
-		"module,__dirname,__filename,exports,require,define",
-
-		// js globals
-		"Array,Date,eval,function,hasOwnProperty,Infinity,isFinite,isNaN,isPrototypeOf,length,Math",
-		"NaN,name,Number,Object,prototype,String,toString,undefined,valueOf",
-
-		// browser globals
-		"alert,all,anchor,anchors,area,assign,blur,button,checkbox,clearInterval,clearTimeout",
-		"clientInformation,close,closed,confirm,constructor,crypto,decodeURI,decodeURIComponent",
-		"defaultStatus,document,element,elements,embed,embeds,encodeURI,encodeURIComponent,escape",
-		"event,fileUpload,focus,form,forms,frame,innerHeight,innerWidth,layer,layers,link,location",
-		"mimeTypes,navigate,navigator,frames,frameRate,hidden,history,image,images,offscreenBuffering",
-		"open,opener,option,outerHeight,outerWidth,packages,pageXOffset,pageYOffset,parent,parseFloat",
-		"parseInt,password,pkcs11,plugin,prompt,propertyIsEnum,radio,reset,screenX,screenY,scroll",
-		"secure,select,self,setInterval,setTimeout,status,submit,taint,text,textarea,top,unescape",
-		"untaint,window",
-
-		// window events
-		"onblur,onclick,onerror,onfocus,onkeydown,onkeypress,onkeyup,onmouseover,onload,onmouseup,onmousedown,onsubmit"
-	]
-		.join(",")
-		.split(",")
-);
-
 /**
  * @template T
  * @param {string} property property
  * @param {function(T[keyof T], T[keyof T]): 0 | 1 | -1} comparator comparator
  * @returns {Comparator<T>} comparator
  */
+
 const createComparator = (property, comparator) => (a, b) =>
 	comparator(
 		a[/** @type {keyof T} */ (property)],
@@ -625,25 +594,6 @@ const getFinalName = (
 					: `/*#__PURE__*/Object(${reference})`;
 		}
 		return reference;
-	}
-};
-
-/**
- * @param {Scope | null} s scope
- * @param {UsedNames} nameSet name set
- * @param {TODO} scopeSet1 scope set 1
- * @param {TODO} scopeSet2 scope set 2
- */
-const addScopeSymbols = (s, nameSet, scopeSet1, scopeSet2) => {
-	let scope = s;
-	while (scope) {
-		if (scopeSet1.has(scope)) break;
-		if (scopeSet2.has(scope)) break;
-		scopeSet1.add(scope);
-		for (const variable of scope.variables) {
-			nameSet.add(variable.name);
-		}
-		scope = scope.upper;
 	}
 };
 
@@ -1193,18 +1143,6 @@ class ConcatenatedModule extends Module {
 		 * @param {string} id export id
 		 * @returns {{ usedNames: UsedNames, alreadyCheckedScopes: Set<TODO> }} info
 		 */
-		const getUsedNamesInScopeInfo = (module, id) => {
-			const key = `${module}-${id}`;
-			let info = usedNamesInScopeInfo.get(key);
-			if (info === undefined) {
-				info = {
-					usedNames: new Set(),
-					alreadyCheckedScopes: new Set()
-				};
-				usedNamesInScopeInfo.set(key, info);
-			}
-			return info;
-		};
 
 		// Set of already checked scopes
 		const ignoredScopes = new Set();
@@ -1274,6 +1212,7 @@ class ConcatenatedModule extends Module {
 							if (!binding.ids) continue;
 							const { usedNames, alreadyCheckedScopes } =
 								getUsedNamesInScopeInfo(
+									usedNamesInScopeInfo,
 									binding.info.module.identifier(),
 									"name" in binding ? binding.name : ""
 								);
@@ -1306,6 +1245,7 @@ class ConcatenatedModule extends Module {
 		// generate names for symbols
 		for (const info of moduleToInfoMap.values()) {
 			const { usedNames: namespaceObjectUsedNames } = getUsedNamesInScopeInfo(
+				usedNamesInScopeInfo,
 				info.module.identifier(),
 				""
 			);
@@ -1314,6 +1254,7 @@ class ConcatenatedModule extends Module {
 					for (const variable of info.moduleScope.variables) {
 						const name = variable.name;
 						const { usedNames, alreadyCheckedScopes } = getUsedNamesInScopeInfo(
+							usedNamesInScopeInfo,
 							info.module.identifier(),
 							name
 						);
@@ -1327,7 +1268,7 @@ class ConcatenatedModule extends Module {
 									ignoredScopes
 								);
 							}
-							const newName = this.findNewName(
+							const newName = findNewName(
 								name,
 								allUsedNames,
 								usedNames,
@@ -1375,7 +1316,7 @@ class ConcatenatedModule extends Module {
 							info.namespaceExportSymbol
 						);
 					} else {
-						namespaceObjectName = this.findNewName(
+						namespaceObjectName = findNewName(
 							"namespaceObject",
 							allUsedNames,
 							namespaceObjectUsedNames,
@@ -1390,7 +1331,7 @@ class ConcatenatedModule extends Module {
 					break;
 				}
 				case "external": {
-					const externalName = this.findNewName(
+					const externalName = findNewName(
 						"",
 						allUsedNames,
 						namespaceObjectUsedNames,
@@ -1404,7 +1345,7 @@ class ConcatenatedModule extends Module {
 			}
 			const buildMeta = /** @type {BuildMeta} */ (info.module.buildMeta);
 			if (buildMeta.exportsType !== "namespace") {
-				const externalNameInterop = this.findNewName(
+				const externalNameInterop = findNewName(
 					"namespaceObject",
 					allUsedNames,
 					namespaceObjectUsedNames,
@@ -1418,7 +1359,7 @@ class ConcatenatedModule extends Module {
 				buildMeta.exportsType === "default" &&
 				buildMeta.defaultObject !== "redirect"
 			) {
-				const externalNameInterop = this.findNewName(
+				const externalNameInterop = findNewName(
 					"namespaceObject2",
 					allUsedNames,
 					namespaceObjectUsedNames,
@@ -1429,7 +1370,7 @@ class ConcatenatedModule extends Module {
 				topLevelDeclarations.add(externalNameInterop);
 			}
 			if (buildMeta.exportsType === "dynamic" || !buildMeta.exportsType) {
-				const externalNameInterop = this.findNewName(
+				const externalNameInterop = findNewName(
 					"default",
 					allUsedNames,
 					namespaceObjectUsedNames,
@@ -1913,53 +1854,6 @@ ${defineGetters}`
 			return ref;
 		});
 		return [list, map];
-	}
-
-	/**
-	 * @param {string} oldName old name
-	 * @param {UsedNames} usedNamed1 used named 1
-	 * @param {UsedNames} usedNamed2 used named 2
-	 * @param {string} extraInfo extra info
-	 * @returns {string} found new name
-	 */
-	findNewName(oldName, usedNamed1, usedNamed2, extraInfo) {
-		let name = oldName;
-
-		if (name === ConcatenationScope.DEFAULT_EXPORT) {
-			name = "";
-		}
-		if (name === ConcatenationScope.NAMESPACE_OBJECT_EXPORT) {
-			name = "namespaceObject";
-		}
-
-		// Remove uncool stuff
-		extraInfo = extraInfo.replace(
-			/\.+\/|(\/index)?\.([a-zA-Z0-9]{1,4})($|\s|\?)|\s*\+\s*\d+\s*modules/g,
-			""
-		);
-
-		const splittedInfo = extraInfo.split("/");
-		while (splittedInfo.length) {
-			name = splittedInfo.pop() + (name ? `_${name}` : "");
-			const nameIdent = Template.toIdentifier(name);
-			if (
-				!usedNamed1.has(nameIdent) &&
-				(!usedNamed2 || !usedNamed2.has(nameIdent))
-			)
-				return nameIdent;
-		}
-
-		let i = 0;
-		let nameWithNumber = Template.toIdentifier(`${name}_${i}`);
-		while (
-			usedNamed1.has(nameWithNumber) ||
-			// eslint-disable-next-line no-unmodified-loop-condition
-			(usedNamed2 && usedNamed2.has(nameWithNumber))
-		) {
-			i++;
-			nameWithNumber = Template.toIdentifier(`${name}_${i}`);
-		}
-		return nameWithNumber;
 	}
 
 	/**
