+++ test/configCases/output-module/iife-entry-module-with-others/module4.js
@@ -1,0 +1,1 @@
+module.exports = typeof Symbol
